\name{conv2}
\alias{conv2}
\title{Two-dimensional Convolution}
\description{This function conducts two-dimensional convolution base on Fourier theory. }
\usage{conv2(X, A)}
\arguments{
  \item{X}{The matrix will be smoothed}
  \item{A}{The smoothing window}
}
\value{Return a smoothed matrix.}
\seealso{\code{\link{WindowBartlett}}}
\author{Huidong Tian}

