\name{FourierRand}
\alias{FourierRand}
\title{Fourier Randomization Simulation}
\description{Generate surrogate time series with similar spectrum feature as the analyzed time series.}
\usage{FourierRand(x)}
\arguments{
  \item{x}{Time series being analyzed}
}
\value{Generate a time series with same length and similar spectrum of x}
\author{Huidong Tian}


