\name{WaveletTransform}
\alias{WaveletTransform}
\title{Wavelet Transform}
\description{Discrete Wavelet Transform}
\usage{WaveletTransform(x, dt = dt, dj = dj, lowerPeriod = lowerPeriod, upperPeriod = upperPeriod)}

\arguments{
  \item{x}{Time series being analyzed}
  \item{dt}{Sampling resolution on temporal field}
  \item{dj}{Sampling resolution on frequency field}
  \item{lowerPeriod}{Lower period of wavelet decomposition}
  \item{upperPeriod}{upper period of wavelet decomposition}
}

\author{Huidong Tian, Bernard Cazelles}

