\name{wc.angle}
\alias{wc.angle}
\title{Plot Wavelet Coherence Phase Shift}
\description{This function adds arrows as the phase shift of wavelet coherence}
\usage{wc.angle(WC = WC, p = 2)}
\arguments{
  \item{WC}{The result of wavelet coherence analysis ("wtc")}
  \item{p}{The region where arrows will be add. }
}
\author{Huidong Tian}
